// Calla/src/resonance-audio/resonance-audio.d.ts
interface ResonanceAudioOptions {
    /**
     * @property {Number} ambisonicOrder
     * Desired ambisonic Order. Defaults to
     * {@linkcode DEFAULT_AMBISONIC_ORDER DEFAULT_AMBISONIC_ORDER}.
     **/
    ambisonicOrder?: number;
    renderingMode?: RenderingMode;
    /**
     * @property {Float32Array} listenerPosition
     * The listener's initial position (in meters), where origin is the center of
     * the room. Defaults to {@linkcode DEFAULT_POSITION DEFAULT_POSITION}.
     **/
    listenerPosition?: vec3;
    /**
     * @property {Float32Array} listenerForward
     * The listener's initial forward vector.
     * Defaults to {@linkcode DEFAULT_FORWARD DEFAULT_FORWARD}.
     **/
    listenerForward?: vec3;
    /**
     * @property {Float32Array} listenerUp
     * The listener's initial up vector.
     * Defaults to {@linkcode DEFAULT_UP DEFAULT_UP}.
     **/
    listenerUp?: vec3;
    /**
     * @property {Utils~RoomDimensions} dimensions Room dimensions (in meters). Defaults to
     * {@linkcode DEFAULT_ROOM_DIMENSIONS DEFAULT_ROOM_DIMENSIONS}.
     **/
    dimensions?: RoomDimensions;
    /**
     * @property {Utils~RoomMaterials} materials Named acoustic materials per wall.
     * Defaults to {@linkcode DEFAULT_ROOM_MATERIALS DEFAULT_ROOM_MATERIALS}.
     **/
    materials?: RoomMaterials;
    /**
     * @property {Number} speedOfSound
     * (in meters/second). Defaults to
     * {@linkcode DEFAULT_SPEED_OF_SOUND DEFAULT_SPEED_OF_SOUND}.
     */
    speedOfSound?: number;
}
/**
 * Main class for managing sources, room and listener models.
 */
export declare class ResonanceAudio implements IDisposable {
    private _sources;
    listener: Listener;
    room: Room;
    ambisonicOrder: number;
    context: AudioContext;
    output: GainNode;
    ambisonicOutput: GainNode;
    ambisonicInput: GainNode;
    /**
     * Main class for managing sources, room and listener models.
     * @param context
     * Associated {@link
    https://developer.mozilla.org/en-US/docs/Web/API/AudioContext AudioContext}.
     * @param options
     * Options for constructing a new ResonanceAudio scene.
     */
    constructor(context: AudioContext, options?: ResonanceAudioOptions);
    getRenderingMode(): string;
    setRenderingMode(mode: string): void;
    dispose(): void;
    /**
     * Create a new source for the scene.
     * @param options
     * Options for constructing a new Source.
     * @return {Source}
     */
    createSource(options: SourceOptions): Source;
    /**
     * Remove an existing source for the scene.
     * @param source
     */
    removeSource(source: Source): void;
    /**
     * Set the scene's desired ambisonic order.
     * @param ambisonicOrder Desired ambisonic order.
     */
    setAmbisonicOrder(ambisonicOrder: number): void;
    /**
     * Set the room's dimensions and wall materials.
     * @param dimensions Room dimensions (in meters).
     * @param materials Named acoustic materials per wall.
     */
    setRoomProperties(dimensions: RoomDimensions, materials: RoomMaterials): void;
    /**
     * Set the listener's position (in meters), where origin is the center of
     * the room.
     */
    setListenerPosition(v: vec3): void;
    /**
     * Set the source's orientation using forward and up vectors.
     */
    setListenerOrientation(forward: vec3, up: vec3): void;
    /**
     * Set the speed of sound.
     */
    setSpeedOfSound(speedOfSound: number): void;
}